send-email: restore --in-reply-to superseding behavior
git send-email --in-reply-to= fails to override the email headers,
if they're present in the output of format-patch, which breaks the
contract of the command line switch. This can cause an email thread
to break, if subsequent replies to a given message need to be sent
after amending fixes to a commit and re-running git format-patch to
get the incremental patch-fix posted.

Signed-off-by: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>